API/requirements.txt
# Unified Platform Dependencies
# ==============================

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# HTTP clients
httpx>=0.26.0
aiohttp>=3.9.0

# Configuration & Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0

# Database
sqlalchemy>=2.0.25
psycopg2-binary>=2.9.9
psycopg-pool>=3.2.0
alembic>=1.13.0

# Redis (caching and rate limiting)
redis>=5.0.0

# LangChain / LangGraph
langchain>=1.0.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langgraph>=0.2.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0
langgraph-checkpoint-postgres>=2.0.0
langsmith>=0.1.0

# LLM APIs
openai>=1.10.0
tiktoken>=0.7.0,<1.0.0

# Authentication
python-jose[cryptography]>=3.3.0
firebase-admin>=6.3.0

# AI/ML APIs
tavily-python>=0.3.0
fal-client>=0.4.0

# Text-to-Speech
edge-tts>=7.0.0
TTS>=0.22.0
soundfile>=0.12.1

# Speech-to-Text (MLX Whisper - local, optimized for Apple Silicon)
mlx-whisper>=0.4.0

# Utilities
fuzzywuzzy>=0.18.0
pypdf>=4.0.0
chromadb>=0.4.0
apscheduler>=3.10.0
parsedatetime>=2.6
pytz>=2024.1
beautifulsoup4>=4.12.0

# Async Support
anyio>=3.0.0

# Development (optional)
pytest>=7.0.0
pytest-asyncio>=0.21.0
black>=23.0.0
ruff>=0.1.0

# Chainlit (optional - for testing UI)
# chainlit>=1.0.0
